requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.6.0
anyio==4.10.0
attrs==25.3.0
black==24.3.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
distlib==0.3.8
exceptiongroup==1.2.0
fake-useragent==1.5.1
Faker==24.4.0
fastapi==0.110.1
fastapi-mcp==0.4.0
filelock==3.13.3
flake8==7.0.0
frozenlist==1.7.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.1
identify==2.5.35
idna==3.6
isort==5.13.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mcp==1.13.1
mdurl==0.1.2
multidict==6.6.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
parsel==1.9.0
pathspec==0.12.1
platformdirs==4.2.0
pre-commit==3.7.0
propcache==0.3.2
pycodestyle==2.11.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.2.0
Pygments==2.17.2
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.1
referencing==0.36.2
requests==2.32.5
rich==13.7.1
rpds-py==0.27.1
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.37.2
tomli==2.2.1
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.25.1
w3lib==2.1.2
watchfiles==0.21.0
websockets==12.0
yarl==1.20.1
